sprucin up details
